refactor(session): realign SessionConfig to agent/session per ADR-056 exit criterion
